// scripts/enums/ServerMessageType.ts
export enum ServerMessageType {
    SET_MY_PLAYER = "set_my_player",
    PLAYER_LIST = "player_list",
    NEW_PLAYER = "new_player",
    PONG = "pong",
    USER_DISCONNECTED = "user_disconnected",
    GROUND = "ground",
    NEW_GROUND = "new_ground",
    FRAME = "frame",
    PLAYER_DEAD = "player_dead",
    USER_LIST = "user_list",
    NEW_USER = "new_user",
    NEW_PROJECTILE = "new_projectile",
    PROJECTILE_INFO = "projectile_info",
    PROJECTILE_DESTROYED = "projectile_destroyed",
    ARRAY = "array",
    PLAYER_INFO = "player_info",
    NEW_SYNCED_OBJECT = "new_synced_object",
    SYNCED_OBJECT_DESTROYED = "synced_object_destroyed",
    SHOP_TIME_LEFT = "shop_time_left",
    NEW_SHOP_TIME = "new_shop_time",
    NEW_LIGHTNING = "new_lightning",
    SPELL_INFO = "spell_info",
    SET_MY_USER = "set_my_user",
    NEW_BOUNCER = "new_bouncer",
    BOUNCER_INFO = "bouncer_info",
    COOLDOWN_INFO = "cooldown_info",
    USER_MONEY = "user_money",
    CAST_START = "cast_start",
    CAST_FINISH = "cast_finish",
    NEW_GAME = "new_game",
    TELEPORT = "teleport",
    NEW_BOOMERANG = "new_boomerang",
    BOOMERANG_INFO = "boomerang_info",
    NEW_SWAP = "new_swap",
    SWAP_INFO = "swap_info",
    NEW_FIREBALL = "new_fireball",
    FIREBALL_INFO = "fireball_info",
    DAMAGE = "damage",
    NEW_BOLDER = "new_bolder",
    BOLDER_INFO = "bolder_info",
    NEW_METEOR = "new_meteor",
    METEOR_INFO = "meteor_info",
    NEW_SPRAY = "new_spray",
    SPRAY_INFO = "spray_info",
    NEW_LINK = "new_link",
    LINK_INFO = "link_info",
    NEW_LINKED = "new_linked",
    LINKED_INFO = "linked_info",
    NEW_GRAVITY = "new_gravity",
    GRAVITY_INFO = "gravity_info",
    NEW_SHIELD = "new_shield",
    SHIELD_DESTROYED = "shield_destroyed",
    NEW_RUSH = "new_rush",
    RUSH_DESTROYED = "rush_destroyed",
    SCOURGE = "scourge",
    PLAY_MODE = "play_mode",
    NEW_COLUMN = "new_column",
    NEW_BARREL = "new_barrel",
    CHAT = "chat",
    PROJECTILE_COLLISION = "projectile_collision",
    PLAYER_STATISTICS = "player_statistics",
    USER_STATISTICS = "user_statistics",
    NEW_HOMING = "new_homing",
    HOMING_INFO = "homing_info",
    CRYSTAL = "crystal",
    ITEM_BOUGHT = "item_bought",
    ITEM_SOLD = "item_sold",
    WITNESS_ME_EXP = "witness_me_exp",
    THRUST_START = "thrust_start",
    THRUST_FINISH = "thrust_finish",
    THRUST_COLLISION = "thrust_collision",
    INNERWARE_USED = "innerware_used",
    SHOP_TIME_FINISHED = "shop_time_finished",
    BUFF_APPLIED = "BUFF_APPLIED",
    BUFF_REMOVED = "BUFF_REMOVED",
    PLAYER_ROUND_RANK = "PLAYER_ROUND_RANK",
    SCORE_TIME = "SCORE_TIME",
    TEAM_SCORE_TIME = "TEAM_SCORE_TIME",
    ROUND_END = "ROUND_END",
    TEAM_ROUND_END = "TEAM_ROUND_END",
    FIGHT_TIME = "FIGHT_TIME",
    ROUND = "round",
    NEW_SPELL = "new_spell",
    SPELL_REMOVED = "spell_removed",
    OBSTACLE_HEALTH = "obstacle_health",
    NEW_OBSTACLE = "new_obstacle",
    PLAYER_SKIN = "player_skin",
    TEAM_ENTRY = "team_entry",
    TEAM_SCORE = "team_score",
    GROUND_CHECKER = "ground_checker",
    DELAYED_EXPLOSION = "delayed_explosion",
    BARREL_BLOW = "barrel_blow",
    OBSTACLE_COLLISION = "obstacle_collision"
}